Add tools.etl.examples.PrintMemcacheStats. Update etl docs with password info.
This example is deliberately very simple: it just makes an RPC and prints the result. This makes a functional test invasive since we have to capture stdout and inspect it. Can simplify this to just run the method and verify it doesn't crash, but that seems like too weak a test.
	Change on 2013/03/26 by johncox <johncox@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=44482227