Allow transforms to operate on KeyProperty instances.
Note that the json representation of a key is a string, so if you convert from entity -> dict -> json -> dict -> entity, the value will be a string rather than a db.Key for all instances after the cast to json. This is consistent with what we do for db.GeoPt: once that hits json, it's a dict rather than a db.GeoPt when restored to dict or entity form.

If we think maintaining key type across the json boundary is important we can add it; I'm just trying to minimize the change in the first pass.
	Change on 2013/06/03 by johncox <johncox@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=47507131