Attempt fix-forward of issues remaining from 48340321:
1) Possible overflow of db.StringProperty 500B limit on unicode input.
2) Provide exactly one update method on StudentProfileDAO that can handle updates of both student and profile instances.
3) Pass the course namespace as an argument, removing the inline import of controllers.sites.
4) Remove early return.

Abhinav, I believe this CL takes care of everything both Pavel and I raised. I've run the tests and they pass, and I've verified the behavior of the app manually (though I admit my understanding of the desired behavior is imperfect).

Apologies in advance if logistically we have to move forward with submit before you get a chance to review -- we need to freeze the codebase so we can hit our release date, and I'm trying to fix forward because the alternative is a rollback, which nobody wants. Should this CL go in, though, you can assume that everything we raised in 48340321 is resolved.
	Change on 2013/06/25 by johncox <johncox@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=48519168