Add notifications module.
Implementation is partial. All objects and interfaces exist; major parts of the implementation are stubbed and will be added in a follow-up. Wanted to get review now to

a) ensure the interfaces provide the functionality you need; and
b) keep the size of the review something approaching humane.
	Change on 2014/05/16 by johncox <johncox@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=67295478