Modify sort ordering so that correct answers to fill-in questions always list first.
Here, the problem is that DC will sort things in a chart by the value returned by
the Dimension.  However, for UI usability, we want to have all incorrect answers
displayed last, and colored differently.  In order to support this, we need to
create a difference between the thing used to get a bar chart label string and the
thing used to determine the bar chart ordering.

To get the label for the bar chart, we return an object whose toString() method
generates the right label.  However, the internal structure of the Answer object
is retained, and this is passed to the ordering() method, which can then have
correct answers sort first.  The sort is stable, so within categories, the answers
remain lexically ordered.
	Change on 2014/07/16 by mgainer <mgainer@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=71279754