Accommodate partial/incomplete fields
I'm noticing an attempt to dereference None at this point when using an older
course.yaml setup.  It's trying to look up 'course' -> 'lti1' -> 'provider_enabled'.
My course has: 'course': { .... 'lti1': ''}
I.e., 'lti1' is present as a key, but not as anything with substructure.
Not sure how it got that way, but the attempt to use .has_key() on a string
is not going as well as you might hope.
	Change on 2014/07/31 by mgainer <mgainer@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=72346173