Add 'core' as pseudo-module so handlers will be shown on Deployment page.
Note that this is a little ugly, but a lot of other fairly core-like things
are configured as modules.  The other option was to do some custom layout
work and add a global to main.py, but that seemed worse than a fake module.
	Change on 2014/09/02 by mgainer <mgainer@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=74607537