Make missing types on upload an error log, not a fatal problem.
Issue here is that as CourseBuilder grows, new types are added.
Until/unless it actually is a fatal error to not have some of
{ResourceBundleEntity, I18nProgressEntity, RoleEntity, LabelEntity},
then not having these should be commented on in the logs, but
should not prevent upload of, say, a course that only has
QuestionEntity and QuestionGroupEntity components.
	Change on 2014/10/06 by mgainer <mgainer@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=77105430