Add support for caching courses larger than 1M in Memcache.
This is a follow-on to the previous work to permit larger courses to be saved
in VFS.  Now that the courses may be present in VFS, we should permit them
to be pre-cached in Memcache.

Note that neither this nor the previous change addresses in-memory caching.
In-memory caching is contemplated directly in the VFS code, as well as
implicitly in the Memcache support code.  Should avoid duplication there.
	Change on 2015/02/19 by mgainer <mgainer@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86742090