Don't collect user_id on course events for non-Student users
For Wipeout, we rely on having an object which is per-Student,
not per-user.  Rather than revisit the entire Wipeout design,
simply don't collect PII for non-Student users.  Instead,
make up a randomized, non-correlatable ID which is persistent
but clear-able and use that instead.  This still gives us the
ability to track a user through multiple pages, but does not
give the ability to match that activity back to any particular
human.  Note that this does also lose the ability to match
pre-registration behavior with post-registration.  This is not
a particularly interesting use-case, so the loss is not terribly
significant.

	Change on 2016/06/29 by mgainer <mgainer@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=126218465