Remove "Pending" state for sitewide config settings.
Note that 'remove' is a bit of a lie; for backward compatibility
with existing settings, we still have the 'is_draft' field so that
we can recognize settings that were marked as Pending prior to
this fix.

Also change behavior: Before this change, a new override entry
was added even before the edit form was shown.  Now we only add
the override upon clicking the Save button on the form.

	Change on 2016/07/01 by mgainer <mgainer@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=126410071