Initial changelist for student_skills_ui module.
This adds a button labeled "Course Map" to the course page. Clicking the button loads a page with a dummy graph containing two nodes and one directed edge that is drawn using D3. (Eventually this graph will be the actual skill map for the course.)

The view of the graph can be shifted by passing x and y parameters in the URL, and scaled by passing a scale parameter.

	Change on 2016/08/04 by tujohnson <tujohnson@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129386369