Post Mortem Debugging.
If you pass the --pdb option to the test runner, the moment a test fails or produces an error it will drop into a debugger.  This feature will make it a lot easier to determine why tests are failing, especially for flakey tests whose failures are difficult to pin down or predict.

	Change on 2016/08/25 by nretallack <nretallack@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=131318333