Fix description on question scoring toggle button
Button toggles between allowing individual scores to be assigned to
different answers in a multiple-choice question.  When in the
mode where individual scores are allowed and the clicking the
button will take us back to a world where correct answers are
worth 1 point and incorrect answers worth 0, label on the button
is currently confusing: "Return to choice-picker".  Change this.

Also small fix so that if individual scores are already set,
text of button includes a warning that toggling back to default
scoring will discard information; this already exists, but is
not set on page load, only after having used the button to
toggle into individual-scores mode.

	Change on 2016/09/26 by mgainer <mgainer@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=134286716