Commit merging multiple CLs.
Moe will not proceed with integrating changes beyond cl/135961950
without manual fixup, which this is.  Problem appears to be related to a
bare reference to a BUILD rule in the internal Moe BUILD file.  For
reasons I don't grasp, using the rule without a corresponding
import statement to bring the rule in was totally fine for several
years, and then breaks on the day of the product release.  Which seems
a little unlikely, admittedly, but here we are.

Main interesting bits for this Git change are cl/135997897 for
fixing multi-edit dialogs and cl/135998062 which suppresses
non-working gcb tags on Explorer pages.  Other CLs are very
small, minor, and mostly related to the imminent release of
1.11

135999237 by mgainer on 2016/10/12 21:44 (local) Add import to BUILD rule
135998883 by mgainer on 2016/10/12 21:37 (local) Record previous use of 'moe' tool
135998558 by mgainer on 2016/10/12 21:31 (local) Add windows config scripts to all_files.txt
135998503 by mgainer on 2016/10/12 21:29 (local) Fix indentation
135998062 by mgainer on 2016/10/12 21:20 (local) Cloned from CL 135965650 by 'g4 patch'.
135997897 by mgainer on 2016/10/12 21:17 (local) Usability and correctness fixes for multi-course edit dialogs
135961950 by mgainer on 2016/10/12 14:25 (local) Update to 2.24 Chromedriver and 2.53.1 Selenium to allow integration tests to work again